Change weather struct to use uint8_t

+++ weatherhelper.h
@@ -2,10 +2,10 @@
 #define _WEATHERHELPER_H_
 
 typedef struct {
-  float temp;
-  float feels_like;
-  float min;
-  float max;
+  uint8_t temp;
+  uint8_t feels_like;
+  uint8_t min;
+  uint8_t max;
   char* desc;
 } Weather;
